Math.Line

Documentation

data Line a

Line in Vector form. A line is a geometrical object in Rⁿ that contains a point P and all points that lie in one direction and its inverse direction from point P.

Every line l in Rⁿ can be parametrized by a vector equation of the form:

l: v➝(t) = v₀➝ + t.d➝	for t ∈

where v₀➝ corresponds to some point on the line and d➝ gives the directio of the line.

Constructors

Line

Fields

point :: Point a

Any point on the line

vector :: Vector a

Direction vector

line :: Point a -> Vector a -> Line a

Creates a line from a point and a Vector

distance :: (Num a, Floating a) => Point a -> Point a -> a

Distance between two points

distPointToLine :: (Num a, Floating a) => Point a -> Line a -> a

Gets the perpendicular distance from a point to a line.

parallelLine :: (Eq a, Floating a) => Line a -> Line a -> Bool

True if two lines are parallel.

perpendicularLine :: (Eq a, Num a) => Line a -> Line a -> Bool

True if two lines are perpendicular