Math.Matrix

Description

A Haskell implementation of Numeric Vectors and Matrices

Documentation

data Matrix a

Constructors

Matrix

Fields

rows :: Int

Number of rows

columns :: Int

Number of Columns

vectors :: Vector (Vector a)

Matrix Creation

matrix

Arguments

:: Int
Number of rows

-> Int
Number of columns

-> ((Int, Int) -> a)
Generator function

-> Matrix a

Creates an m x n matrix with the given generator function.

identity :: Num a => Int -> Matrix a

zeroMatrix :: Num a => Int -> Int -> Matrix a

Creates an m x n matrix of 0's

inOrderMatrix :: Int -> Int -> Matrix Int

Accessing

Accessing elements

getElem :: (Int, Int) -> Matrix a -> a

returns the element at position (i,j)

row :: Int -> Matrix a -> Vector a

returns the row at the the given index. Note: index starts at 0.

column :: Int -> Matrix a -> Vector a

returns the row at the the given index. Note: index starts at 0.

mainDiagonal :: Matrix a -> Vector a

Returns the main diagonal

Properties

size :: Matrix a -> (Int, Int)

Returns the size of the matrix as a pair (rows, columns)

square :: Matrix a -> Bool

returns True if matrix is square

Matrix Manipulation

Single Matrix Operations

transpose :: Matrix a -> Matrix a

Returns the transpose of a matrix Swaps the rows and columns of a vector

multiplyMatrix :: Num a => Matrix a -> Matrix a -> Matrix a

Synonym for (*) Multiplies the two given matrices. Creates a new matrix from the dot product of the rows of m and the columns of n

mZipWith :: (a -> b -> c) -> Matrix a -> Matrix b -> Matrix c

zipWith function defined for matrices

Row and Column Operations

mapRow :: (a -> a) -> Int -> Matrix a -> Matrix a

mapColumn :: (a -> a) -> Int -> Matrix a -> Matrix a

swapRow :: Int -> Int -> Matrix a -> Matrix a

swapColumn :: Int -> Int -> Matrix a -> Matrix a

Joining Matrices

(<|>) :: Matrix a -> Matrix a -> Matrix a

(<->) :: Matrix a -> Matrix a -> Matrix a

Conversions

toLists :: Matrix a -> [[a]]

fromLists :: [[a]] -> Matrix a

Converts lists to Matrix

>>> fromLists [[1,2,3], [4,5,6]]
| 1 2 3 |
| 4 5 6 |

fromVectors :: Vector (Vector a) -> Matrix a

toVector :: Matrix a -> Vector a

Creates a vector from a matrix In the order (0,0), (0,1), .. (0,n), (1,0), (1,1), .. (n,n)