Math.Vector

Description

A Haskell implementation of Numeric Vectors and Matrices

This class defines an extension of Vector to introduce numeric operation for vectors. It begins by adding an instance for Num Vector, allowing for numeric vectors to be treated as numbers.

example: let u = 1,2,3 let v = 3,4,5 then u + v = 4,6,8

It also implements common vector functions such as dot and (eventually) cross products.

Documentation

type Point = Vector

type Vector = Vector

Operators

(<.>) :: Num a => Vector a -> Vector a -> a

Infix operator for the dot product of two Vectors

(-|) :: (Eq a, Num a) => Vector a -> Vector a -> Bool

Infix operator for perpendicular

(-||) :: (Eq a, Floating a) => Vector a -> Vector a -> Bool

Infix operator for parallel

(*>) :: (Functor f, Num a) => f a -> a -> f a

Multiply a functor (Matrix or Vector) by a scalar

(<*) :: (Functor f, Num a) => a -> f a -> f a

Multiply a functor (Matrix or Vector) by a scalar

Vector Creation

newVector :: Num a => Point a -> Point a -> Vector a

Vector a relative to b

readVector :: Read a => String -> Vector a

Reads a vector from a String in the form x1, .. xn

Vector Operations

Angles

angle :: Floating a => Vector a -> Vector a -> a

Get the angle between two vectors

parallel :: (Eq a, Floating a) => Vector a -> Vector a -> Bool

True when two vectors are parallel May have trouble with rounding errors

perpendicular :: (Eq a, Num a) => Vector a -> Vector a -> Bool

True when two vectors are parallel

Dot and cross products

dot :: Num a => Vector a -> Vector a -> a

Short form of dotProduct

dotProduct :: Num a => Vector a -> Vector a -> a

Dot product of two Vectors

projection :: Floating a => Vector a -> Vector a -> Vector a

Projection of u onto v

Single vector operations

magnitude :: Floating a => Vector a -> a

Returns the magnitude / vector norm of a vector

unitVector :: Floating a => Vector a -> Vector a

returns the unit vector of u

pointToVector :: Point a -> Vector a

Converts a Point to a Vector