Math.Matrix

Description

A Haskell implementation of Numeric Vectors, Matrices and Lines.

Documentation

data Matrix a

Matrix implementation using the Vector class. A matrix is defined as a Vector of Vectors. This implementation allows for Matrices of any size from 1 * 1 up to intMax * intMax, (theoretically but don't try it).

>>> print $ identity 4
| 1 0 0 0 |
| 0 1 0 0 |
| 0 0 1 0 |
| 0 0 0 1 |

Constructors

Matrix

Fields

rows :: Int

Number of rows

columns :: Int

Number of Columns

vectors :: Vector (Vector a)

Matrix Creation

matrix

Arguments

:: Int
Number of rows

-> Int
Number of columns

-> ((Int, Int) -> a)
Generator function

-> Matrix a

Creates an m x n matrix with the given generator function.

The generator function can be any funtion of type (Int,Int) -> a. Where (Int, Int) represents the position (x,y) in the matrix and a is the result at that position.

Example:

\(x,y) -> x + y

>>> matrix 4 4 (\(x,y) -> x + y)
| 0 1 2 3 |  which is equivalent to: | (0 + 0), (0 + 1), (0 + 2), (0 + 3) |
| 1 2 3 4 |                          | (1 + 0), (1 + 1), (1 + 2), (1 + 3) |
| 2 3 4 5 |                          | (2 + 0), (2 + 1), (2 + 2), (2 + 3) |
| 3 4 5 6 |                          | (3 + 0), (3 + 1), (3 + 2), (3 + 3) |

identity :: Num a => Int -> Matrix a

Creates an n x n matrix where the main diagonal = 0.

>>> identity 3
| 1 0 0 |
| 0 1 0 |
| 0 0 1 |

zeroMatrix :: Num a => Int -> Int -> Matrix a

Creates an m x n matrix of 0's

inOrderMatrix :: Int -> Int -> Matrix Int

Creates an n x m matrix of Ints in ascending order.

>>> inOrderMatrix 4 4
|  1  2  3  4 |
|  5  6  7  8 |
|  9 10 11 12 |
| 13 14 15 16 |

enumMatrix

Arguments

:: Enum a

=> Int
Rows

-> Int
Columns

-> a
Starting value

-> Matrix a

Similar to inOrderMatrix but is given a starting value. The starting value must be an enumeratable type; Int, Char, etc.

>>> enumMatrix 3 3 'a'
| 'a' 'b' 'c' |
| 'd' 'e' 'f' |
| 'g' 'h' 'i' |

Accessing

Accessing elements

getElem :: (Int, Int) -> Matrix a -> a

returns the element at position (i,j)

row :: Int -> Matrix a -> Vector a

returns the row at the the given index. Note: index starts at 0.

column :: Int -> Matrix a -> Vector a

returns the row at the the given index. Note: index starts at 0.

mainDiagonal :: Matrix a -> Vector a

Returns the main diagonal

Properties

size :: Matrix a -> (Int, Int)

Returns the size of the matrix as a pair (rows, columns)

square :: Matrix a -> Bool

returns True if matrix is square

Matrix Manipulation

Single Matrix Operations

transpose :: Matrix a -> Matrix a

Returns the transpose of a matrix Swaps the rows and columns of a vector

multiplyMatrix :: Num a => Matrix a -> Matrix a -> Matrix a

Synonym for (*) Multiplies the two given matrices. Creates a new matrix from the dot product of the rows of m and the columns of n

mZipWith :: (a -> b -> c) -> Matrix a -> Matrix b -> Matrix c

zipWith function defined for matrices

Row and Column Operations

mapRow :: (a -> a) -> Int -> Matrix a -> Matrix a

Maps a function onto a row

mapColumn :: (a -> a) -> Int -> Matrix a -> Matrix a

Maps a function onto a column

swapRow :: Int -> Int -> Matrix a -> Matrix a

Swap two rows

swapColumn :: Int -> Int -> Matrix a -> Matrix a

Swap two columns

Joining Matrices

(<|>) :: Matrix a -> Matrix a -> Matrix a

(<->) :: Matrix a -> Matrix a -> Matrix a

Conversions

toLists :: Matrix a -> [[a]]

Converts the matrix to a list of lists.

fromList :: Int -> Int -> [a] -> Matrix a

Converts a single list into a matrix of the given size. The list must contain enough elements to fill the matrix.

fromLists :: [[a]] -> Matrix a

Converts lists to Matrix

>>> fromLists [[1,2,3], [4,5,6]]
| 1 2 3 |
| 4 5 6 |

fromVectors :: Vector (Vector a) -> Matrix a

Creates a matrix from a Vector of Vectors

toVector :: Matrix a -> Vector a

Creates a vector from a matrix In the order (0,0), (0,1), .. (0,n), (1,0), (1,1), .. (n,n)